+++ pyserini/2cr/msmarco.py
@@ -26,9 +26,200 @@ from string import Template
 import pkg_resources
 import yaml
 
-from ._base import run_eval_and_return_metric, ok_str, okish_str, fail_str, \
-    find_msmarco_table_topic_set_key_v1, find_msmarco_table_topic_set_key_v2
-from .defs_msmarco import models, trec_eval_metric_definitions
+from ._base import run_eval_and_return_metric, ok_str, okish_str, fail_str
+
+# The models: the rows of the results table will be ordered this way.
+models = {
+    'msmarco-v1-passage':
+    ['bm25-default',
+     'bm25-rm3-default',
+     'bm25-rocchio-default',
+     '',
+     'bm25-tuned',
+     'bm25-rm3-tuned',
+     'bm25-rocchio-tuned',
+     '',
+     'bm25-d2q-t5-default',
+     'bm25-rm3-d2q-t5-default',
+     'bm25-rocchio-d2q-t5-default',
+     '',
+     'bm25-d2q-t5-tuned',
+     'bm25-rm3-d2q-t5-tuned',
+     'bm25-rocchio-d2q-t5-tuned',
+     '',
+     'unicoil-noexp',
+     'unicoil',
+     '',
+     'unicoil-noexp-otf',
+     'unicoil-otf',
+     '',
+     'ance',
+     'distilbert-kd',
+     'distilbert-kd-tasb',
+     'tct_colbert-v2-hnp',
+     '',
+     'ance-otf',
+     'distilbert-kd-otf',
+     'distilbert-kd-tasb-otf',
+     'tct_colbert-v2-hnp-otf'],
+    'msmarco-v1-doc':
+    ['bm25-doc-default',
+     'bm25-doc-segmented-default',
+     'bm25-rm3-doc-default',
+     'bm25-rm3-doc-segmented-default',
+     'bm25-rocchio-doc-default',
+     'bm25-rocchio-doc-segmented-default',
+     '',
+     'bm25-doc-tuned',
+     'bm25-doc-segmented-tuned',
+     'bm25-rm3-doc-tuned',
+     'bm25-rm3-doc-segmented-tuned',
+     'bm25-rocchio-doc-tuned',
+     'bm25-rocchio-doc-segmented-tuned',
+     '',
+     'bm25-d2q-t5-doc-default',
+     'bm25-d2q-t5-doc-segmented-default',
+     'bm25-rm3-d2q-t5-doc-default',
+     'bm25-rm3-d2q-t5-doc-segmented-default',
+     '',
+     'bm25-d2q-t5-doc-tuned',
+     'bm25-d2q-t5-doc-segmented-tuned',
+     'bm25-rm3-d2q-t5-doc-tuned',
+     'bm25-rm3-d2q-t5-doc-segmented-tuned',
+     '',
+     'unicoil-noexp',
+     'unicoil',
+     '',
+     'unicoil-noexp-otf',
+     'unicoil-otf'],
+    'msmarco-v2-passage':
+    ['bm25-default',
+     'bm25-augmented-default',
+     'bm25-rm3-default',
+     'bm25-rm3-augmented-default',
+     '',
+     'bm25-d2q-t5-default',
+     'bm25-d2q-t5-augmented-default',
+     'bm25-rm3-d2q-t5-default',
+     'bm25-rm3-d2q-t5-augmented-default',
+     '',
+     'unicoil-noexp',
+     'unicoil',
+     '',
+     'unicoil-noexp-otf',
+     'unicoil-otf'],
+    'msmarco-v2-doc':
+    ['bm25-doc-default',
+     'bm25-doc-segmented-default',
+     'bm25-rm3-doc-default',
+     'bm25-rm3-doc-segmented-default',
+     '',
+     'bm25-d2q-t5-doc-default',
+     'bm25-d2q-t5-doc-segmented-default',
+     'bm25-rm3-d2q-t5-doc-default',
+     'bm25-rm3-d2q-t5-doc-segmented-default',
+     '',
+     'unicoil-noexp',
+     'unicoil',
+     '',
+     'unicoil-noexp-otf',
+     'unicoil-otf'
+     ]
+}
+
+trec_eval_metric_definitions = {
+    'msmarco-v1-passage': {
+        'msmarco-passage-dev-subset': {
+            'MRR@10': '-c -M 10 -m recip_rank',
+            'R@1K': '-c -m recall.1000'
+        },
+        'dl19-passage': {
+            'MAP': '-c -l 2 -m map',
+            'nDCG@10': '-c -m ndcg_cut.10',
+            'R@1K': '-c -l 2 -m recall.1000'
+        },
+        'dl20-passage': {
+            'MAP': '-c -l 2 -m map',
+            'nDCG@10': '-c -m ndcg_cut.10',
+            'R@1K': '-c -l 2 -m recall.1000'
+        }
+    },
+    'msmarco-v1-doc': {
+        'msmarco-doc-dev': {
+            'MRR@10': '-c -M 100 -m recip_rank',
+            'R@1K': '-c -m recall.1000'
+        },
+        'dl19-doc': {
+            'MAP': '-c -M 100 -m map',
+            'nDCG@10': '-c -m ndcg_cut.10',
+            'R@1K': '-c -m recall.1000'
+        },
+        'dl20-doc': {
+            'MAP': '-c -M 100 -m map',
+            'nDCG@10': '-c -m ndcg_cut.10',
+            'R@1K': '-c -m recall.1000'
+        }
+    },
+    'msmarco-v2-passage': {
+        'msmarco-v2-passage-dev': {
+            'MRR@100': '-c -M 100 -m recip_rank',
+            'R@1K': '-c -m recall.1000'
+        },
+        'msmarco-v2-passage-dev2': {
+            'MRR@100': '-c -M 100 -m recip_rank',
+            'R@1K': '-c -m recall.1000'
+        },
+        'dl21-passage': {
+            'MAP@100': '-c -l 2 -M 100 -m map',
+            'nDCG@10': '-c -m ndcg_cut.10',
+            'MRR@100': '-c -l 2 -M 100 -m recip_rank',
+            'R@100': '-c -l 2 -m recall.100',
+            'R@1K': '-c -l 2 -m recall.1000'
+        }
+    },
+    'msmarco-v2-doc': {
+        'msmarco-v2-doc-dev': {
+            'MRR@100': '-c -M 100 -m recip_rank',
+            'R@1K': '-c -m recall.1000'
+        },
+        'msmarco-v2-doc-dev2': {
+            'MRR@100': '-c -M 100 -m recip_rank',
+            'R@1K': '-c -m recall.1000'
+        },
+        'dl21-doc': {
+            'MAP@100': '-c -M 100 -m map',
+            'nDCG@10': '-c -m ndcg_cut.10',
+            'MRR@100': '-c -M 100 -m recip_rank',
+            'R@100': '-c -m recall.100',
+            'R@1K': '-c -m recall.1000'
+        }
+    }
+}
+
+
+def find_msmarco_table_topic_set_key_v1(topic_key):
+    # E.g., we want to map variants like 'dl19-passage-unicoil' and 'dl19-passage' both into 'dl19'
+    key = ''
+    if topic_key.startswith('dl19'):
+        key = 'dl19'
+    elif topic_key.startswith('dl20'):
+        key = 'dl20'
+    elif topic_key.startswith('msmarco'):
+        key = 'dev'
+
+    return key
+
+
+def find_msmarco_table_topic_set_key_v2(topic_key):
+    key = ''
+    if topic_key.endswith('dev') or topic_key.endswith('dev-unicoil') or topic_key.endswith('dev-unicoil-noexp'):
+        key = 'dev'
+    elif topic_key.endswith('dev2') or topic_key.endswith('dev2-unicoil') or topic_key.endswith('dev2-unicoil-noexp'):
+        key = 'dev2'
+    elif topic_key.startswith('dl21'):
+        key = 'dl21'
+
+    return key
 
 
 def format_command(raw):
@@ -47,27 +238,27 @@ def read_file(f):
     return text
 
 
+def list_conditions(args):
+    print(models[args.collection])
+
+
 def generate_report(args):
-    if args.collection == 'v1-passage':
-        collection = 'msmarco-v1-passage'
+    if args.collection == 'msmarco-v1-passage':
         yaml_file = pkg_resources.resource_filename(__name__, 'msmarco-v1-passage.yaml')
         html_template = read_file(pkg_resources.resource_filename(__name__, 'msmarco_html_v1_passage.template'))
         row_template = read_file(pkg_resources.resource_filename(__name__, 'msmarco_html_row_v1.template'))
-    elif args.collection == 'v1-doc':
-        collection = 'msmarco-v1-doc'
-        yaml_file = 'pyserini/resources/msmarco-v1-doc.yaml'
-        html_template = read_file('scripts/repro_matrix/msmarco_html_v1_doc.template')
-        row_template = read_file('scripts/repro_matrix/msmarco_html_row_v1.template')
-    elif args.collection == 'v2-passage':
-        collection = 'msmarco-v2-passage'
-        yaml_file = 'pyserini/resources/msmarco-v2-passage.yaml'
-        html_template = read_file('scripts/repro_matrix/msmarco_html_v2.template')
-        row_template = read_file('scripts/repro_matrix/msmarco_html_row_v2.template')
-    elif args.collection == 'v2-doc':
-        collection = 'msmarco-v2-doc'
-        yaml_file = 'pyserini/resources/msmarco-v2-doc.yaml'
-        html_template = read_file('scripts/repro_matrix/msmarco_html_v2.template')
-        row_template = read_file('scripts/repro_matrix/msmarco_html_row_v2.template')
+    elif args.collection == 'msmarco-v1-doc':
+        yaml_file = pkg_resources.resource_filename(__name__, 'msmarco-v1-doc.yaml')
+        html_template = read_file(pkg_resources.resource_filename(__name__, 'msmarco_html_v1_doc.template'))
+        row_template = read_file(pkg_resources.resource_filename(__name__, 'msmarco_html_row_v1.template'))
+    elif args.collection == 'msmarco-v2-passage':
+        yaml_file = pkg_resources.resource_filename(__name__, 'msmarco-v2-passage.yaml')
+        html_template = read_file(pkg_resources.resource_filename(__name__, 'msmarco_html_v2.template'))
+        row_template = read_file(pkg_resources.resource_filename(__name__, 'msmarco_html_row_v2.template'))
+    elif args.collection == 'msmarco-v2-doc':
+        yaml_file = pkg_resources.resource_filename(__name__, 'msmarco-v2-doc.yaml')
+        html_template = read_file(pkg_resources.resource_filename(__name__, 'msmarco_html_v2.template'))
+        row_template = read_file(pkg_resources.resource_filename(__name__, 'msmarco_html_row_v2.template'))
     else:
         raise ValueError(f'Unknown corpus: {args.collection}')
 
@@ -94,27 +285,27 @@ def generate_report(args):
                 eval_key = topic_set['eval_key']
 
                 short_topic_key = ''
-                if collection == 'msmarco-v1-passage' or collection == 'msmarco-v1-doc':
+                if args.collection == 'msmarco-v1-passage' or args.collection == 'msmarco-v1-doc':
                     short_topic_key = find_msmarco_table_topic_set_key_v1(topic_key)
                 else:
                     short_topic_key = find_msmarco_table_topic_set_key_v2(topic_key)
 
-                runfile = f'run.{collection}.{name}.{short_topic_key}.txt'
+                runfile = f'run.{args.collection}.{name}.{short_topic_key}.txt'
                 cmd = Template(cmd_template).substitute(topics=topic_key, output=runfile)
                 commands[name][short_topic_key] = cmd
 
                 for expected in topic_set['scores']:
                     for metric in expected:
                         eval_cmd = f'python -m pyserini.eval.trec_eval ' + \
-                                   f'{trec_eval_metric_definitions[collection][eval_key][metric]} {eval_key} {runfile}'
+                                   f'{trec_eval_metric_definitions[args.collection][eval_key][metric]} {eval_key} {runfile}'
                         eval_commands[name][short_topic_key] += eval_cmd + '\n'
                         table[name][short_topic_key][metric] = expected[metric]
 
-    if collection == 'msmarco-v1-passage' or collection == 'msmarco-v1-doc':
+    if args.collection == 'msmarco-v1-passage' or args.collection == 'msmarco-v1-doc':
         row_cnt = 1
 
         html_rows = []
-        for name in models[collection]:
+        for name in models[args.collection]:
             if not name:
                 # Add blank row for spacing
                 html_rows.append('<tr><td style="border-bottom: 0"></td></tr>')
@@ -141,17 +332,23 @@ def generate_report(args):
 
             # If we don't have scores, we want to remove the commands also. Use simple regexp substitution.
             if table[name]['dl19']['MAP'] == 0:
-                s = re.sub(re.compile('Command to generate run on TREC 2019 queries:.*?</div>', re.MULTILINE | re.DOTALL), 'Not available.</div>', s)
+                s = re.sub(re.compile('Command to generate run on TREC 2019 queries:.*?</div>',
+                                      re.MULTILINE | re.DOTALL),
+                           'Not available.</div>', s)
             if table[name]['dl20']['MAP'] == 0:
-                s = re.sub(re.compile('Command to generate run on TREC 2020 queries:.*?</div>', re.MULTILINE | re.DOTALL), 'Not available.</div>', s)
+                s = re.sub(re.compile('Command to generate run on TREC 2020 queries:.*?</div>',
+                                      re.MULTILINE | re.DOTALL),
+                           'Not available.</div>', s)
             if table[name]['dev']['MRR@10'] == 0:
-                s = re.sub(re.compile('Command to generate run on dev queries:.*?</div>', re.MULTILINE | re.DOTALL), 'Not available.</div>', s)
+                s = re.sub(re.compile('Command to generate run on dev queries:.*?</div>',
+                                      re.MULTILINE | re.DOTALL),
+                           'Not available.</div>', s)
 
             html_rows.append(s)
             row_cnt += 1
 
         all_rows = '\n'.join(html_rows)
-        if collection == 'msmarco-v1-passage':
+        if args.collection == 'msmarco-v1-passage':
             full_name = 'MS MARCO V1 Passage'
         else:
             full_name = 'MS MARCO V1 Document'
@@ -162,7 +359,7 @@ def generate_report(args):
         row_cnt = 1
 
         html_rows = []
-        for name in models[collection]:
+        for name in models[args.collection]:
             if not name:
                 # Add blank row for spacing
                 html_rows.append('<tr><td style="border-bottom: 0"></td></tr>')
@@ -191,7 +388,7 @@ def generate_report(args):
             row_cnt += 1
 
         all_rows = '\n'.join(html_rows)
-        if collection == 'msmarco-v2-passage':
+        if args.collection == 'msmarco-v2-passage':
             full_name = 'MS MARCO V2 Passage'
         else:
             full_name = 'MS MARCO V2 Document'
@@ -206,20 +403,7 @@ def run_conditions(args):
     table = defaultdict(lambda: defaultdict(lambda: defaultdict(lambda: 0.0)))
     table_keys = {}
 
-    if args.collection == 'v1-passage':
-        collection = 'msmarco-v1-passage'
-        yaml_file = pkg_resources.resource_filename(__name__, 'msmarco-v1-passage.yaml')
-    elif args.collection == 'v1-doc':
-        collection = 'msmarco-v1-doc'
-        yaml_file = 'pyserini/resources/msmarco-v1-doc.yaml'
-    elif args.collection == 'v2-passage':
-        collection = 'msmarco-v2-passage'
-        yaml_file = 'pyserini/resources/msmarco-v2-passage.yaml'
-    elif args.collection == 'v2-doc':
-        collection = 'msmarco-v2-doc'
-        yaml_file = 'pyserini/resources/msmarco-v2-doc.yaml'
-    else:
-        raise ValueError(f'Unknown corpus: {args.collection}')
+    yaml_file = pkg_resources.resource_filename(__name__, f'{args.collection}.yaml')
 
     with open(yaml_file) as f:
         yaml_data = yaml.safe_load(f)
@@ -239,14 +423,14 @@ def run_conditions(args):
                 eval_key = topic_set['eval_key']
 
                 short_topic_key = ''
-                if collection == 'msmarco-v1-passage' or collection == 'msmarco-v1-doc':
+                if args.collection == 'msmarco-v1-passage' or args.collection == 'msmarco-v1-doc':
                     short_topic_key = find_msmarco_table_topic_set_key_v1(topic_key)
                 else:
                     short_topic_key = find_msmarco_table_topic_set_key_v2(topic_key)
 
                 print(f'  - topic_key: {topic_key}')
 
-                runfile = os.path.join(args.directory, f'run.{collection}.{name}.{short_topic_key}.txt')
+                runfile = os.path.join(args.directory, f'run.{args.collection}.{name}.{short_topic_key}.txt')
                 cmd = Template(cmd_template).substitute(topics=topic_key, output=runfile)
 
                 if args.display_commands:
@@ -266,10 +450,11 @@ def run_conditions(args):
                                 continue
 
                             score = float(
-                                run_eval_and_return_metric(metric,
-                                                           eval_key,
-                                                           trec_eval_metric_definitions[collection][eval_key][metric],
-                                                           runfile))
+                                run_eval_and_return_metric(
+                                    metric,
+                                    eval_key,
+                                    trec_eval_metric_definitions[args.collection][eval_key][metric],
+                                    runfile))
                             if math.isclose(score, float(expected[metric])):
                                 result_str = ok_str
                             # Flaky test: small difference on my iMac Studio
@@ -286,7 +471,7 @@ def run_conditions(args):
                 if not args.skip_eval:
                     print('')
 
-    if collection == 'msmarco-v1-passage' or collection == 'msmarco-v1-doc':
+    if args.collection == 'msmarco-v1-passage' or args.collection == 'msmarco-v1-doc':
         print(' ' * 69 + 'TREC 2019' + ' ' * 16 + 'TREC 2020' + ' ' * 12 + 'MS MARCO dev')
         print(' ' * 62 + 'MAP    nDCG@10    R@1K       MAP nDCG@10    R@1K    MRR@10    R@1K')
         print(' ' * 62 + '-' * 22 + '    ' + '-' * 22 + '    ' + '-' * 14)
@@ -300,7 +485,7 @@ def run_conditions(args):
                   f'{table[name]["dev"]["MRR@10"]:8.4f}{table[name]["dev"]["R@1K"]:8.4f}')
         else:
             # Otherwise, print out all rows
-            for name in models[collection]:
+            for name in models[args.collection]:
                 if not name:
                     print('')
                     continue
@@ -312,7 +497,7 @@ def run_conditions(args):
         print(' ' * 77 + 'TREC 2021' + ' ' * 18 + 'MS MARCO dev' + ' ' * 6 + 'MS MARCO dev2')
         print(' ' * 62 + 'MAP@100 nDCG@10 MRR@100 R@100   R@1K     MRR@100   R@1K    MRR@100   R@1K')
         print(' ' * 62 + '-' * 38 + '    ' + '-' * 14 + '    ' + '-' * 14)
-        for name in models[collection]:
+        for name in models[args.collection]:
             if not name:
                 print('')
                 continue
@@ -330,18 +515,36 @@ def run_conditions(args):
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(description='Generate regression matrix for MS MARCO corpora.')
-    parser.add_argument('--collection', type=str, help='Collection = {v1-passage, v1-doc, v2-passage, v2-doc}.', required=True)
+    parser.add_argument('--collection', type=str,
+                        help='Collection = {v1-passage, v1-doc, v2-passage, v2-doc}.', required=True)
+    # To list all conditions
+    parser.add_argument('--list-conditions', action='store_true', default=False, help='List available conditions.')
     # For generating reports
     parser.add_argument('--generate-report', action='store_true', default=False, help='Generate report.')
-    parser.add_argument('--output', type=str, help='Output', required=False)
+    parser.add_argument('--output', type=str, help='File to store report.', required=False)
     # For actually running the experimental conditions
-    parser.add_argument('--all', action='store_true', default=False, help='Run all commands.')
-    parser.add_argument('--condition', type=str, help='Condition', required=False)
-    parser.add_argument('--directory', type=str, help='Base directory', default='', required=False)
-    parser.add_argument('--dry-run', action='store_true', default=False, help='Print out command but not execute.')
+    parser.add_argument('--all', action='store_true', default=False, help='Run all conditions.')
+    parser.add_argument('--condition', type=str, help='Condition to run.', required=False)
+    parser.add_argument('--directory', type=str, help='Base directory.', default='', required=False)
+    parser.add_argument('--dry-run', action='store_true', default=False, help='Print out commands but do not execute.')
     parser.add_argument('--skip-eval', action='store_true', default=False, help='Skip running trec_eval.')
     parser.add_argument('--display-commands', action='store_true', default=False, help='Display command.')
     args = parser.parse_args()
+
+    if args.collection == 'v1-passage':
+        args.collection = 'msmarco-v1-passage'
+    elif args.collection == 'v1-doc':
+        args.collection = 'msmarco-v1-doc'
+    elif args.collection == 'v2-passage':
+        args.collection = 'msmarco-v2-passage'
+    elif args.collection == 'v2-doc':
+        args.collection = 'msmarco-v2-doc'
+    else:
+        raise ValueError(f'Unknown corpus: {args.collection}')
+
+    if args.list_conditions:
+        list_conditions(args)
+        sys.exit()
 
     if args.generate_report:
         if not args.output:
